+++ documentation/plugins/docgen.ts
@@ -212,7 +212,9 @@ const createParser = (configPath: string) => {
                 prop.declarations.length > 0 &&
                 !Boolean(prop.declarations.find(declarationFilter));
 
-            if (isExcluded || isExternal) {
+            const isUnknown = typeof prop.declarations === "undefined";
+
+            if (isExcluded || isExternal || isUnknown) {
                 return false;
             }
             return true;