+++ packages/mantine/src/components/fields/date/index.tsx
@@ -8,6 +8,8 @@ import { Text, TextProps } from "@mantine/core";
 
 dayjs.extend(LocalizedFormat);
 
+const defaultLocale = dayjs.locale();
+
 export type DateFieldProps = RefineFieldDateProps<ConfigType, TextProps>;
 
 /**
@@ -18,8 +20,15 @@ export type DateFieldProps = RefineFieldDateProps<ConfigType, TextProps>;
  */
 export const DateField: React.FC<DateFieldProps> = ({
     value,
+    locales,
     format: dateFormat = "L",
     ...rest
 }) => {
-    return <Text {...rest}>{dayjs(value).format(dateFormat)}</Text>;
+    return (
+        <Text {...rest}>
+            {dayjs(value)
+                .locale(locales || defaultLocale)
+                .format(dateFormat)}
+        </Text>
+    );
 };